refactor: rename project detail response
